scm-plugins/scm-hg-plugin/src/test/java/sonia/scm/repository/spi/HgCommandContextFactoryTest.java
/*
 * Copyright (c) 2020 - present Cloudogu GmbH
 *
 * This program is free software: you can redistribute it and/or modify it under
 * the terms of the GNU Affero General Public License as published by the Free
 * Software Foundation, version 3.
 *
 * This program is distributed in the hope that it will be useful, but WITHOUT
 * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
 * FOR A PARTICULAR PURPOSE. See the GNU Affero General Public License for more
 * details.
 *
 * You should have received a copy of the GNU Affero General Public License
 * along with this program. If not, see https://www.gnu.org/licenses/.
 */

package sonia.scm.repository.spi;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import sonia.scm.repository.HgConfigResolver;
import sonia.scm.repository.HgRepositoryFactory;
import sonia.scm.repository.RepositoryTestData;

import static org.assertj.core.api.Assertions.assertThat;

@ExtendWith(MockitoExtension.class)
class HgCommandContextFactoryTest {

  @Mock
  private HgConfigResolver configResolver;

  @Mock
  private HgRepositoryFactory repositoryFactory;

  @InjectMocks
  private HgCommandContextFactory commandContextFactory;

  @Test
  void shouldCreateHgCommandContext() {
    HgCommandContext hg = commandContextFactory.create(RepositoryTestData.createHeartOfGold("hg"));
    assertThat(hg).isNotNull();
  }

}
